LinUCB + Simulator Edit
Implementation of LinUSB policy
Simulator: parameters from file, dataset parsing changed

+++ src/exploChallenge/logs/GenericVisitor.java
@@ -3,9 +3,9 @@ package exploChallenge.logs;
 public class GenericVisitor {
 
 	private long timestamp;
-	private byte[] features;
+	private double[] features;
  
-	public GenericVisitor(long timestamp, byte[] features) {
+	public GenericVisitor(long timestamp, double[] features) {
 		this.features = features;
 		this.timestamp = timestamp;
 	}
@@ -14,7 +14,7 @@ public class GenericVisitor {
 		return timestamp;
 	}
 
-	public byte[] getFeatures() {
+	public double[] getFeatures() {
 		return features;
 	}
